add protections

+++ Cpp_01/ex01/zombie.hpp
@@ -13,4 +13,4 @@ class	Zombie
 		void	announce() const;
 };
 
-Zombie	*zombieHorde(int N, std::string name);
+Zombie	*zombieHorde(size_t N, std::string name);
